+++ internal/adapter/src/repository.rs
@@ -169,7 +169,7 @@ mod tests {
     use kernel::error::{AggregateError, ApplyCommandError, LoadEventError};
     use kernel::processor::CommandProcessor;
     use kernel::Id;
-    use lib::Error;
+    use lib::{DateTime, Error};
     use testcontainers::clients::Cli;
     use testcontainers_modules::mysql::Mysql;
 
@@ -187,46 +187,6 @@ mod tests {
 
     const WIDGET_NAME: &str = "部品名";
     const WIDGET_DESCRIPTION: &str = "部品の説明";
-
-    #[allow(dead_code)]
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-    enum DateTime {
-        /// 2023-01-01T00:00:00Z
-        DT2023_01_01_00_00_00_00,
-        /// 2023-01-01T00:00:01Z
-        DT2023_01_01_00_00_00_01,
-        /// 2023-01-01T00:00:02Z
-        DT2023_01_01_00_00_00_02,
-        /// 2023-01-01T00:01:00Z
-        DT2023_01_01_00_00_01_00,
-        /// 2023-01-01T01:00:00Z
-        DT2023_01_01_00_01_00_00,
-        /// 2023-01-01T01:00:00Z
-        DT2023_01_01_01_00_00_00,
-        /// 2023-01-02T00:00:00Z
-        DT2023_01_02_00_00_00_00,
-        /// 2023-02-01T00:00:00Z
-        DT2023_02_01_00_00_00_00,
-        /// 2024-01-01T00:00:00Z
-        DT2024_01_01_00_00_00_00,
-    }
-
-    impl DateTime {
-        fn id(self) -> String {
-            match self {
-                DateTime::DT2023_01_01_00_00_00_00 => "01GNNA1J00PQ9J874NBWERBM3Z",
-                DateTime::DT2023_01_01_00_00_00_01 => "01GNNA1J015CFH0CA590B4K9K6",
-                DateTime::DT2023_01_01_00_00_00_02 => "01GNNA1J02N9H1YCMRA2R9Q562",
-                DateTime::DT2023_01_01_00_00_01_00 => "01GNNA1JZ86A6F1G8HV7NYHDCN",
-                DateTime::DT2023_01_01_00_01_00_00 => "01GNNA3CK0B63HH8HBYQVRJ5Y8",
-                DateTime::DT2023_01_01_01_00_00_00 => "01GNNDFDM0WV3PR6RM8TEA7MZ5",
-                DateTime::DT2023_01_02_00_00_00_00 => "01GNQWE9003DQHKPAAHCDCVTJZ",
-                DateTime::DT2023_02_01_00_00_00_00 => "01GR57SPM0XBGEG4A13ZBW02G2",
-                DateTime::DT2024_01_01_00_00_00_00 => "01HK153X00D14NM09FKYEJ7MPY",
-            }
-            .to_string()
-        }
-    }
 
     #[derive(Debug, Clone, PartialEq, Eq)]
     enum Fixture {
