+++ internal/app/Cargo.toml
@@ -12,4 +12,5 @@ thiserror = "1.0.56"
 
 [dev-dependencies]
 kernel = { path = "../kernel", features = ["test", "mockall"] }
+lib = { path = "../lib", features = ["test"] }
 tokio = { version = "1.36.0", default-features = false, features = ["macros", "rt-multi-thread"] }
